nhat_ky_xe_files/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pytz==2024.1
google-api-python-client==2.140.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pandas==2.2.2
openpyxl==3.1.5
